requirements.txt
torch==2.1.2
torchvision==0.16.2
opencv-python==4.7.*
mlflow==2.9.2
tqdm==4.65.0
lightning==2.1.3
pydantic==1.10.9
retinaface_pytorch
transformers==4.36.2
evaluate==0.4.1